git-svn-id: file:///home/gecchi/repos@3653 9df8dbd1-7660-4e89-b1e7-fa117e7ea0c7

+++ GgafDx9Core/src/jp/ggaf/dx9core/sound/GgafDx9Bgm.cpp
@@ -31,7 +31,7 @@ GgafDx9Bgm::GgafDx9Bgm(char* prm_bgm_key) : GgafObject() {
     pPcmPlayer = NEW PCMPlayer(GgafDx9Sound::_pIDirectSound8 , pOggDecoder);
 }
 
-void GgafDx9Bgm::play(int prm_volume, int prm_pan, bool prm_isLoop) {
+void GgafDx9Bgm::play(int prm_volume, float prm_pan, bool prm_isLoop) {
     setVolume(prm_volume);
     setPan(prm_pan);
     pPcmPlayer->play(prm_isLoop);
@@ -46,11 +46,13 @@ void GgafDx9Bgm::stop() {
 }
 
 void GgafDx9Bgm::setVolume(int prm_volume) {
+    //ボリューム→デシベル
     int db = GgafDx9Sound::aDbVolume[(LONG)(prm_volume * GgafDx9Sound::_master_volume_rate * GgafDx9Sound::_bgm_volume_rate)];
     pPcmPlayer->setVolume(db);
 }
 
-void GgafDx9Bgm::setPan(int prm_pan) {
+void GgafDx9Bgm::setPan(float prm_pan) {
+
     pPcmPlayer->setPan(prm_pan*DSBPAN_RIGHT);
 }
